refactor: consolidate utility functions
